refactor: common-domain 이름 common-module로 변경
