+++ src/backend/apigateway-server/build.gradle
@@ -30,7 +30,7 @@ ext {
 
 dependencies {
     // 공통 모듈 추가
-    implementation project(":common-domain")
+    implementation project(":common-module")
 
     // Spring
     implementation 'org.springframework.boot:spring-boot-starter-web'
