--- src/backend/common-domain/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'common-domain'
