--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/CommonResponse.java
@@ -1,32 +1,0 @@
-package com.bbebig.commondomain.response.code;
-
-import com.bbebig.commondomain.response.code.success.BaseCode;
-import com.bbebig.commondomain.response.code.success.SuccessStatus;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-@JsonPropertyOrder({"code", "message", "result"})
-public class CommonResponse<T> {
-
-	private final String code;
-	private final String message;
-	@JsonInclude(JsonInclude.Include.NON_NULL)
-	private T result;
-
-
-	public static <T> CommonResponse<T> onSuccess(T result) {
-		return new CommonResponse<>(SuccessStatus._OK.getCode(), SuccessStatus._OK.getMessage(), result);
-	}
-
-	public static <T> CommonResponse<T> of(BaseCode code, T result) {
-		return new CommonResponse<>(code.getReasonHttpStatus().getCode(), code.getReasonHttpStatus().getMessage(), result);
-	}
-
-	public static <T> CommonResponse<T> onFailure(String code, String message, T data) {
-		return new CommonResponse<>(code, message, data);
-	}
-}
