--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/error/BaseErrorCode.java
@@ -1,8 +1,0 @@
-package com.bbebig.commondomain.response.code.error;
-
-public interface BaseErrorCode {
-
-    public ErrorReasonDTO getReason();
-
-    public ErrorReasonDTO getReasonHttpStatus();
-}
