--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/error/ErrorReasonDTO.java
@@ -1,14 +1,0 @@
-package com.bbebig.commondomain.response.code.error;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ErrorReasonDTO<T> {
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-}
