--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/success/BaseCode.java
@@ -1,9 +1,0 @@
-package com.bbebig.commondomain.response.code.success;
-
-public interface BaseCode {
-
-    public ReasonDTO getReason();
-
-    public ReasonDTO getReasonHttpStatus();
-
-}
