--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/success/ReasonDTO.java
@@ -1,14 +1,0 @@
-package com.bbebig.commondomain.response.code.success;
-
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-public class ReasonDTO {
-    private HttpStatus httpStatus;
-
-    private final String code;
-    private final String message;
-}
