--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/code/success/SuccessStatus.java
@@ -1,33 +1,0 @@
-package com.bbebig.commondomain.response.code.success;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum SuccessStatus implements BaseCode {
-    _OK(HttpStatus.OK, "200", "성공입니다."),
-    ;
-
-    private final HttpStatus httpStatus;
-    private final String code;
-    private final String message;
-
-    @Override
-    public ReasonDTO getReason() {
-        return ReasonDTO.builder()
-            .code(code)
-            .message(message)
-            .build();
-    }
-
-    @Override
-    public ReasonDTO getReasonHttpStatus() {
-        return ReasonDTO.builder()
-            .httpStatus(httpStatus)
-            .code(code)
-            .message(message)
-            .build();
-    }
-}
