--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/exception/BaseExceptionType.java
@@ -1,16 +1,0 @@
-package com.bbebig.commondomain.response.exception;
-
-import org.springframework.http.HttpStatus;
-
-/**
- * 인터페이스 BaseExceptionType
- * 예외의 세부 정보를 제공하는 메서드를 정의.
- */
-public interface BaseExceptionType {
-
-    int getErrorCode();
-
-    HttpStatus getHttpStatus();
-
-    String getErrorMessage();
-}
