--- src/backend/common-domain/src/main/java/com/bbebig/commondomain/response/exception/ErrorHandler.java
@@ -1,14 +1,0 @@
-package com.bbebig.commondomain.response.exception;
-
-import com.bbebig.commondomain.response.code.error.BaseErrorCode;
-
-/**
- * 클래스 ErrorHandler
- * 특정 오류 코드를 기반으로 예외를 처리하는 클래스
- */
-public class ErrorHandler extends GeneralException {
-
-    public ErrorHandler(BaseErrorCode errorCode) {
-        super(errorCode);
-    }
-}
