--- src/backend/common-domain/src/main/proto/Passport.proto
@@ -1,14 +1,0 @@
-syntax = "proto3";
-
-package com.bbebig.common;
-
-option java_package = "com.bbebig.common.proto";
-option java_outer_classname = "PassportProto";
-
-message Passport {
-  string passport_id = 1;
-  int64 member_id = 2;
-  int64 issued_at = 3;
-  int64 expires_at = 4;
-  string auth_level = 5;
-}
