+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/code/error/BaseErrorCode.java
@@ -1,0 +1,8 @@
+package com.bbebig.commonmodule.response.code.error;
+
+public interface BaseErrorCode {
+
+    public ErrorReasonDTO getReason();
+
+    public ErrorReasonDTO getReasonHttpStatus();
+}
