+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/code/error/ErrorReasonDTO.java
@@ -1,0 +1,14 @@
+package com.bbebig.commonmodule.response.code.error;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ErrorReasonDTO<T> {
+
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+}
