+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/code/success/BaseCode.java
@@ -1,0 +1,9 @@
+package com.bbebig.commonmodule.response.code.success;
+
+public interface BaseCode {
+
+    public ReasonDTO getReason();
+
+    public ReasonDTO getReasonHttpStatus();
+
+}
