+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/code/success/ReasonDTO.java
@@ -1,0 +1,14 @@
+package com.bbebig.commonmodule.response.code.success;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ReasonDTO {
+    private HttpStatus httpStatus;
+
+    private final String code;
+    private final String message;
+}
