+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/code/success/SuccessStatus.java
@@ -1,0 +1,33 @@
+package com.bbebig.commonmodule.response.code.success;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum SuccessStatus implements BaseCode {
+    _OK(HttpStatus.OK, "200", "성공입니다."),
+    ;
+
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+
+    @Override
+    public ReasonDTO getReason() {
+        return ReasonDTO.builder()
+            .code(code)
+            .message(message)
+            .build();
+    }
+
+    @Override
+    public ReasonDTO getReasonHttpStatus() {
+        return ReasonDTO.builder()
+            .httpStatus(httpStatus)
+            .code(code)
+            .message(message)
+            .build();
+    }
+}
