+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/exception/BaseException.java
@@ -1,0 +1,11 @@
+package com.bbebig.commonmodule.response.exception;
+
+/**
+ * 추상 클래스 BaseException
+ * 모든 사용자 정의 예외의 기본 클래스
+ * getExceptionType 메서드를 구현
+ */
+public abstract class BaseException extends RuntimeException {
+
+    public abstract BaseExceptionType getExceptionType();
+}
