+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/response/exception/GeneralException.java
@@ -1,0 +1,25 @@
+package com.bbebig.commonmodule.response.exception;
+
+import com.bbebig.commonmodule.response.code.error.BaseErrorCode;
+import com.bbebig.commonmodule.response.code.error.ErrorReasonDTO;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * 클래스 GeneralException
+ * 기본 예외 처리를 위한 클래스
+ */
+@Getter
+@AllArgsConstructor
+public class GeneralException extends RuntimeException {
+
+    private final BaseErrorCode code;
+
+    public ErrorReasonDTO getErrorReason() {
+        return this.code.getReason();
+    }
+
+    public ErrorReasonDTO getErrorReasonHttpStatus() {
+        return this.code.getReasonHttpStatus();
+    }
+}
